requirements.txt
# Core dependencies
python-dotenv>=1.0.0
pyyaml>=6.0.1

# Audio processing
pyaudio>=0.2.13
webrtcvad>=2.0.10
pvporcupine>=2.2.0

# Speech services
edge-tts>=6.1.9
openai>=1.3.0

# Async support
aiohttp>=3.9.1
asyncio>=3.4.3

# Testing
pytest>=7.4.3
pytest-asyncio>=0.23.2
